Fixed an issue where relying on SE summary amounts would ignore the monthly reset until AFTER  the first stream of the month.
We're now using SE goal data to compare local data instead of SE summary data from their API.

Of note, this will now check and reset the data on Sunday if you have weekly setup for StreamElements, and on the first of the month it will reset your goals on SE automagically.

As a result of this, if you do a SE reset using the Advanced Goal Manager it will correctly reflect and fully reset instead of still using the summary amounts as before.

This is a bit more flexible now in regards to data from SE.